update design of ProjectDetailPage
