+++ lib/screens/projects/project_square.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:guam_front/screens/projects/project_detail.dart';
+import 'package:guam_front/screens/projects/detail/project_detail.dart';
 import '../../models/project.dart';
 import 'package:hexcolor/hexcolor.dart';
